requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2
databases
asyncpg
passlib[bcrypt]
python-dotenv
fastapi-login
pyjwt
